exercise_09 version2

+++ ani.sadaghyan/repo/Lesson_3/exercise_09/exercise_09.py
@@ -1,4 +1,4 @@
-meal = int(input("Write price of the meal: "))
+meal = float(input("Write price of the meal: "))
 percent  = int(input("Percent tip: "))
 tip = meal * percent / 100
 print ("Your price for tip: ", tip)
